Move to Java and improve tests
